Kernel/Devices: Move ConsoleDevice into the Devices source directory
