+++ Kernel/CMakeLists.txt
@@ -42,11 +42,11 @@ set(KERNEL_SOURCES
     Bus/VirtIO/RNG.cpp
     CMOS.cpp
     CommandLine.cpp
-    ConsoleDevice.cpp
     Coredump.cpp
     Devices/AsyncDeviceRequest.cpp
     Devices/BlockDevice.cpp
     Devices/CharacterDevice.cpp
+    Devices/ConsoleDevice.cpp
     Devices/Device.cpp
     Devices/DeviceManagement.cpp
     Devices/FullDevice.cpp
